Move `lib/gen*.c` to `tools/gen*.cpp` 	修改：     .gitignore 	修改：     ChangeLog 	修改：     Makefile.devel 	删除：     lib/genaliases.c 	删除：     lib/genaliases2.c 	修改：     lib/generated/flags.h 	删除：     lib/genflags.c 	新文件：   tools/genaliases.cpp 	新文件：   tools/genaliases2.cpp 	新文件：   tools/genflags.cpp
